Refactor analytics to ditch Apollo and accommodate async noop script load
