+++ package.json
@@ -12,7 +12,6 @@
     "chromatic": "chromatic --project-token w4dua5t94wh"
   },
   "dependencies": {
-    "@apollo/client": "^3.3.6",
     "@rent_avail/base": "^0.5.0",
     "@rent_avail/controls": "^0.2.16",
     "@rent_avail/input": "^0.4.5",
@@ -20,13 +19,11 @@
     "@rent_avail/menu": "^0.2.8",
     "@rent_avail/typography": "^0.1.20",
     "@rent_avail/utils": "^0.5.0",
-    "apollo-link-prismic": "^1.0.8",
     "clsx": "^1.1.1",
     "framer-motion": "^3.1.1",
     "fromentries": "^1.3.2",
     "graphql": "^15.4.0",
     "jsonwebtoken": "^8.5.1",
-    "lru-cache": "^6.0.0",
     "next": "^10.2.2",
     "next-compose-plugins": "^2.2.1",
     "next-seo": "^4.24.0",
