+++ pages/_app.js
@@ -5,16 +5,12 @@ import PreviewWarning from "components/partials/PreviewWarning"
 import { DefaultSeo } from "next-seo"
 import { BREAKPOINTS, DEFAULT_SEO } from "config"
 import { UIDReset } from "react-uid"
-import { ApolloClient, ApolloProvider, InMemoryCache } from "@apollo/client"
-import resolvers from "graphql/resolvers"
-import AnalyticsGlobals from "components/partials/AnalyticsGlobals"
+import { AnalyticsProvider } from "utils/analytics/context"
 
 const marketingTheme = {
   ...theme,
   breakpoints: [...BREAKPOINTS],
 }
-
-const apolloClient = new ApolloClient({ cache: new InMemoryCache(), resolvers })
 
 const GlobalStyles = createGlobalStyle`
  .async-hide { opacity: 0 !important}
@@ -24,8 +20,7 @@ export default function App({ Component, pageProps }) {
   const { preview } = pageProps
   return (
     <ThemeProvider theme={marketingTheme}>
-      <ApolloProvider client={apolloClient}>
-        <AnalyticsGlobals />
+      <AnalyticsProvider>
         <DefaultSeo {...DEFAULT_SEO} />
         <Reset />
         <GlobalStyles />
@@ -33,7 +28,7 @@ export default function App({ Component, pageProps }) {
         <UIDReset prefix="uid_">
           <Component {...pageProps} />
         </UIDReset>
-      </ApolloProvider>
+      </AnalyticsProvider>
     </ThemeProvider>
   )
 }
