+++ pages/_document.js
@@ -33,8 +33,6 @@ export default class Document extends NextDoc {
   render() {
     const GTM_CONTAINER_ID = false
     const SEGMENT_KEY = false
-    const AVAIL_BASE_URL =
-      process.env.NEXT_PUBLIC_AVAIL_BASE_URL || "https://www.avail.co"
     return (
       <Html lang="en-US">
         <Head>
@@ -83,10 +81,6 @@ export default class Document extends NextDoc {
               }}
             />
           )}
-          <script
-            async
-            src={`${AVAIL_BASE_URL}/api/v2/public/analytics/noop.js`}
-          />
           <link
             href="https://fonts.googleapis.com/css2?family=Nunito:wght@200;400;600;800&display=swap"
             rel="stylesheet"
