--- src/graphql/queries/index.js
@@ -1,9 +1,0 @@
-import { gql } from "@apollo/client"
-
-export const GET_ANALYTICS_QUERY = gql`
-  query getAnalytics {
-    analytics @client {
-      uuid
-    }
-  }
-`
